Move most classes to jupiter package
