+++ src/main/java/org/junitpioneer/jupiter/Issue.java
@@ -10,12 +10,12 @@
 
 package org.junitpioneer.jupiter;
 
-import org.junit.jupiter.api.extension.ExtendWith;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
  * {@code @Issue} is a JUnit Jupiter extension to mark tests that
