+++ src/main/java/org/junitpioneer/jupiter/IssueExtension.java
@@ -10,11 +10,11 @@
 
 package org.junitpioneer.jupiter;
 
+import java.util.Optional;
+
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junitpioneer.jupiter.issue.IssueExtensionExecutionListener;
-
-import java.util.Optional;
 
 /**
  * This class implements the functionality for the {@code @Issue} annotation.
@@ -27,24 +27,24 @@ class IssueExtension implements BeforeEachCallback {
 	public void beforeEach(ExtensionContext context) {
 		Optional<Issue> annotation = PioneerAnnotationUtils.findClosestEnclosingAnnotation(context, Issue.class);
 
-		if(annotation.isPresent()) {
+		if (annotation.isPresent()) {
 			String issueId = annotation.get().value();
 			context.publishReportEntry(IssueExtensionExecutionListener.REPORT_ENTRY_KEY, issueId);
 		}
 
 	}
 
-//	/**
-//	 * Reads the {@code @Issue} value from the annotation.
-//	 * @param context The Extensions context
-//	 * @return The read value
-//	 */
-//	String readIssueIdFromAnnotation(ExtensionContext context) {
-//		return AnnotationSupport
-//				.findAnnotation(context.getElement(), Issue.class)
-//				.map(Issue::value)
-//				.orElseThrow(() -> new ExtensionConfigurationException(
-//					"The extension is active, but the corresponding annotation could not be found. (This may be a bug.)"));
-//	}
+	//	/**
+	//	 * Reads the {@code @Issue} value from the annotation.
+	//	 * @param context The Extensions context
+	//	 * @return The read value
+	//	 */
+	//	String readIssueIdFromAnnotation(ExtensionContext context) {
+	//		return AnnotationSupport
+	//				.findAnnotation(context.getElement(), Issue.class)
+	//				.map(Issue::value)
+	//				.orElseThrow(() -> new ExtensionConfigurationException(
+	//					"The extension is active, but the corresponding annotation could not be found. (This may be a bug.)"));
+	//	}
 
 }
