+++ src/main/java/org/junitpioneer/jupiter/IssuedTestCase.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junitpioneer.jupiter.issue;
+package org.junitpioneer.jupiter;
 
 /**
  * Represents the execution result of test method, which is annotated with {@link org.junitpioneer.jupiter.Issue}.
